+++ src/Containers/Interfaces/IStackInventory.cs
@@ -70,13 +70,6 @@ namespace TheChest.Inventories.Containers.Interfaces
 
         #region Add
         /// <summary>
-        /// Checks if <paramref name="item"/> can be added to any slot on inventory.
-        /// </summary>
-        /// <param name="item">The item to evaluate to add to the inventory.</param>
-        /// <returns>true if the <paramref name="item"/> can be added; otherwise, false.</returns>
-        [Obsolete("Use CanAdd(params T[]) instead. This method will be removed in future versions.")]
-        bool CanAdd(T item);
-        /// <summary>
         /// Checks if <paramref name="items"/> can be added to any slot on inventory.
         /// </summary>
         /// <param name="items">An array of items to evaluate for addition to the inventory.</param>
@@ -106,14 +99,6 @@ namespace TheChest.Inventories.Containers.Interfaces
         #endregion
 
         #region AddAt
-        /// <summary>
-        /// Determines whether the specified item can be added at the given index.
-        /// </summary>
-        /// <param name="item">The item to evaluate for insertion at the specified index.</param>
-        /// <param name="index">The zero-based index at which to check if the item can be added.</param>
-        /// <returns><see langword="true"/> if the item can be added at the specified index; otherwise, <see langword="false"/>.</returns>
-        [Obsolete("Use CanAddAt(T[], int) instead. This method will be removed in future versions.")]
-        bool CanAddAt(T item, int index);
         /// <summary>
         /// Determines whether the specified items can be added at the given index.
         /// </summary>